feat(tui): right-align the branch and give /effort a picker
Two things the status row and the effort command got wrong for a reader.

The branch sat between the effort and the workspace, so it moved every
time the model name or level changed length and had to be hunted for.
Right-align it at the far edge, where it holds one column; a row with no
room left drops it rather than truncating, because half a branch name
reads as a different branch.

`/effort` only accepted a level typed as an argument, which means knowing
the levels before asking for them. A bare `/effort` now opens a numbered
picker in the shape of the model picker — one row per level plus `off`,
marked at the current setting, answered by number, by name, or by an
empty line to keep what is set. `/effort high` still sets it outright.

+++ crates/arsy-kernel/tests/provider.rs
@@ -402,7 +402,7 @@ fn effort_becomes_a_thinking_budget_inside_the_output_budget() {
 
     // The floor fitting is not the same as an answer fitting: a budget must
     // leave at least as much room to answer as it takes to think.
-    let smallest_with_thinking = (THINKING_FLOOR * 2) as u32;
+    let smallest_with_thinking = THINKING_FLOOR * 2;
     for (max_tokens, expected) in [
         (smallest_with_thinking - 1, None),
         (smallest_with_thinking, Some(THINKING_FLOOR)),
